Remove the old dropdown
