--- packages/insomnia/src/ui/components/base/dropdown/dropdown-divider.tsx
@@ -1,14 +1,0 @@
-import classnames from 'classnames';
-import React, { FC, ReactNode } from 'react';
-interface Props {
-  children?: ReactNode;
-}
-export const DropdownDivider: FC<Props> = ({ children }) => (
-  <div
-    className={classnames('dropdown__divider', {
-      'dropdown__divider--no-name': !children,
-    })}
-  >
-    <span className="dropdown__divider__label">{children}</span>
-  </div>
-);
